requirements.txt
cassandra-driver
flask
python-dotenv
